refactor(datetime): extract shared @shelf/datetime formatter package
Move the webapp's pure, locale-leak-free date/time formatter into a new
`@shelf/datetime` workspace package so the webapp and the companion (Expo/RN)
app share ONE implementation and their date rendering never drifts.

- New `packages/datetime` (TS source, consumed like `@shelf/database`): the
  formatter core verbatim, with the two webapp-coupled type imports replaced by
  local string-union enums (mirroring the Prisma enums member-for-member) and a
  dependency-free `FormatHints` type in place of `ClientHint`.
- `apps/webapp/app/utils/date-format.ts` becomes a thin `export *` re-export, so
  all ~54 webapp import sites are unchanged. Zero behavior change — the webapp's
  57 formatter tests pass against the re-exported package.
- Wire the package into both bundlers (webapp `ssr.noExternal`; Metro transpiles
  the TS source directly, verified via an export build).
