+++ apps/companion/package.json
@@ -30,6 +30,7 @@
     "@react-navigation/bottom-tabs": "^7.2.0",
     "@react-navigation/native": "^7.0.14",
     "@sentry/react-native": "~7.2.0",
+    "@shelf/datetime": "workspace:*",
     "@shelf/labels": "workspace:*",
     "@supabase/supabase-js": "^2.49.1",
     "expo": "~54.0.33",
