+++ packages/datetime/tsconfig.json
@@ -1,0 +1,4 @@
+{
+  "extends": "@shelf/typescript-config/base.json",
+  "include": ["src/**/*.ts"]
+}
